[refactor] Builder周りを整理
