+++ lib/src/commonMain/kotlin/com/wsr/layers/function/ReluD1.kt
@@ -25,6 +25,4 @@ class ReluD1 internal constructor(override val numOfInput: Int) : Layer.D1() {
     }
 }
 
-fun Network.Builder.reluD1() = addLayer(ReluD1(numOfInput = numOfInput))
-
 fun <T : IOType> NetworkBuilder.D1<T>.relu() = addLayer(ReluD1(numOfInput = numOfInput))
