+++ lib/src/commonMain/kotlin/com/wsr/layers/function/SoftmaxD1.kt
@@ -24,7 +24,4 @@ class SoftmaxD1 internal constructor(override val numOfInput: Int) : Layer.D1() 
     }
 }
 
-fun Network.Builder.softmaxD1() =
-    addLayer(SoftmaxD1(numOfInput = numOfInput))
-
 fun <T : IOType> NetworkBuilder.D1<T>.softmax() = addLayer(SoftmaxD1(numOfInput = numOfInput))
